[Lotto] refactor : exception 패키지 분리
