+++ src/test/java/lotto/domain/lottoTicket/LottoNumberTest.java
@@ -1,6 +1,6 @@
 package lotto.domain.lottoTicket;
 
-import lotto.util.InvalidLottoNumberException;
+import lotto.exception.InvalidLottoNumberException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
